+++ xknx/io/routing.py
@@ -6,8 +6,11 @@ Routing uses UDP Multicast to broadcast and receive KNX/IP messages.
 from __future__ import annotations
 
 import asyncio
+from collections.abc import AsyncIterator
+from contextlib import asynccontextmanager
 import logging
-from typing import TYPE_CHECKING
+import random
+from typing import TYPE_CHECKING, Final
 
 from xknx.core import XknxConnectionState
 from xknx.exceptions import CommunicationError
@@ -17,18 +20,110 @@ from xknx.knxip import (
     CEMIMessageCode,
     KNXIPFrame,
     KNXIPServiceType,
+    RoutingBusy,
     RoutingIndication,
+    RoutingLostMessage,
 )
-from xknx.telegram import TelegramDirection
+from xknx.telegram import Telegram, TelegramDirection
 
 from .interface import Interface, TelegramCallbackType
 from .transport import KNXIPTransport, UDPTransport
 
 if TYPE_CHECKING:
-    from xknx.telegram import Telegram
     from xknx.xknx import XKNX
 
 logger = logging.getLogger("xknx.log")
+
+BUSY_DECREMENT_TIME: Final = 0.005  # 5 ms
+BUSY_INCREMENT_COOLDOWN: Final = 0.01  # 10 ms
+BUSY_RANDOM_TIME_FACTOR: Final = 0.05  # 50 ms
+BUSY_SLOWDURATION_TIME_FACTOR: Final = 0.1  # 100 ms
+ROUTING_INDICATION_WAIT_TIME: Final = 0.02  # 20 ms
+
+
+class _RoutingFlowControl:
+    """
+    Class for hanling KNX/IP routing flow control.
+
+    See KNX Specifications 3.8.5 Routing §2.3.5 Flow control handling
+    """
+
+    def __init__(self) -> None:
+        self._last_busy_frame_time: float = 0.0
+        self._last_sent_routing_indication_time: float = 0.0
+        self._loop = asyncio.get_running_loop()
+        self._ready = asyncio.Event()
+        self._ready.set()
+        self._received_busy_frames: int = 0
+        self._timer_task: asyncio.Task[None] | None = None
+        self._wait_start_time: float | None = None
+        self._wait_time_ms: int = 0
+
+    def cancel(self) -> None:
+        """Cancel internal tasks."""
+        if self._timer_task:
+            self._timer_task.cancel()
+
+    @asynccontextmanager
+    async def throttle(self) -> AsyncIterator[None]:
+        """Context manager to wait for ready state and throttle outgoing frames."""
+        # limit RoutingIndication transmission rate according to
+        # KNX Specifications 3.2.6 Communication Medium KNX IP §2.1
+        # simplified version - pause 20 ms after transmit a RoutingIndication
+        elapsed = self._loop.time() - self._last_sent_routing_indication_time
+        if elapsed < ROUTING_INDICATION_WAIT_TIME:
+            await asyncio.sleep(ROUTING_INDICATION_WAIT_TIME - elapsed)
+
+        await self._ready.wait()
+        yield
+        self._last_sent_routing_indication_time = self._loop.time()
+
+    def handle_routing_busy(self, routing_busy: RoutingBusy) -> None:
+        """Handle incoming RoutingBusy."""
+        self._ready.clear()
+        now = self._loop.time()
+        previous_busy_frame_time = self._last_busy_frame_time
+        self._last_busy_frame_time = now
+        if self._wait_start_time is None:
+            logger.info(
+                "RoutingBusy received: %s",
+                routing_busy,
+            )
+        else:
+            # only apply if we have already received a RoutingBusy frame and are still pausing
+            if (now - previous_busy_frame_time) > BUSY_INCREMENT_COOLDOWN:
+                self._received_busy_frames += 1
+            logger.debug(
+                "RoutingBusy received: %s - %s ms since previous - number %s in moving time window",
+                routing_busy,
+                round((now - previous_busy_frame_time) * 1000),
+                self._received_busy_frames,
+            )
+            # discard frame if wait time is lower than remaining time
+            remaining_ms = (now - self._wait_start_time) * 1000
+            if remaining_ms >= routing_busy.wait_time:
+                return
+        self._wait_time_ms = routing_busy.wait_time
+        self._wait_start_time = now
+
+        if self._timer_task:
+            self._timer_task.cancel()
+        self._timer_task = asyncio.create_task(self._resume_sending())
+
+    async def _resume_sending(self) -> None:
+        """Reset ready flag after wait_time_ms and fade out slowduration."""
+        random_wait_extension = (
+            random.random() * self._received_busy_frames * BUSY_RANDOM_TIME_FACTOR
+        )
+        slowduration = self._received_busy_frames * BUSY_SLOWDURATION_TIME_FACTOR
+        await asyncio.sleep(self._wait_time_ms / 1000 + random_wait_extension)
+
+        self._ready.set()
+        self._wait_start_time = None
+        await asyncio.sleep(slowduration)
+        while self._received_busy_frames > 0:
+            await asyncio.sleep(BUSY_DECREMENT_TIME)
+            self._received_busy_frames -= 1
 
 
 class Routing(Interface):
@@ -50,48 +145,21 @@ class Routing(Interface):
             remote_addr=(self.xknx.multicast_group, self.xknx.multicast_port),
             multicast=True,
         )
-
         self.udp_transport.register_callback(
-            self.response_rec_callback, [KNXIPServiceType.ROUTING_INDICATION]
+            self._handle_frame,
+            [
+                KNXIPServiceType.ROUTING_INDICATION,
+                KNXIPServiceType.ROUTING_BUSY,
+                KNXIPServiceType.ROUTING_LOST_MESSAGE,
+            ],
         )
+        self._flow_control = _RoutingFlowControl()
 
-    def response_rec_callback(
-        self, knxipframe: KNXIPFrame, source: HPAI, _: KNXIPTransport
-    ) -> None:
-        """Verify and handle knxipframe. Callback from internal udp_transport."""
-        if not isinstance(knxipframe.body, RoutingIndication):
-            logger.warning("Service type not implemented: %s", knxipframe)
-        elif knxipframe.body.cemi is None:
-            # ignore unsupported CEMI frame
-            return
-        elif knxipframe.body.cemi.src_addr == self.xknx.own_address:
-            logger.debug("Ignoring own packet")
-        else:
-            # TODO: is cemi message code L_DATA.req or .con valid for routing? if not maybe warn and ignore
-            asyncio.create_task(self.handle_cemi_frame(knxipframe.body.cemi))
-
-    async def handle_cemi_frame(self, cemi: CEMIFrame) -> None:
-        """Handle incoming telegram and send responses if applicable (device management)."""
-        telegram = cemi.telegram
-        telegram.direction = TelegramDirection.INCOMING
-
-        if response_tgs := await self.telegram_received_callback(telegram):
-            for response in response_tgs:
-                await self.send_telegram(response)
-
-    async def send_telegram(self, telegram: "Telegram") -> None:
-        """Send Telegram to routing connected device."""
-        cemi = CEMIFrame.init_from_telegram(
-            telegram=telegram,
-            code=CEMIMessageCode.L_DATA_IND,
-            src_addr=self.xknx.own_address,
-        )
-        routing_indication = RoutingIndication(cemi=cemi)
-        await self.send_knxipframe(KNXIPFrame.init_from_body(routing_indication))
-
-    async def send_knxipframe(self, knxipframe: KNXIPFrame) -> None:
-        """Send KNXIPFrame to connected routing device."""
-        self.udp_transport.send(knxipframe)
+    ####################
+    #
+    # CONNECT DISCONNECT
+    #
+    ####################
 
     async def connect(self) -> bool:
         """Start routing."""
@@ -123,3 +191,70 @@ class Routing(Interface):
         await self.xknx.connection_manager.connection_state_changed(
             XknxConnectionState.DISCONNECTED
         )
+        self._flow_control.cancel()
+
+    ##################
+    #
+    # OUTGOING FRAMES
+    #
+    ##################
+
+    async def send_telegram(self, telegram: Telegram) -> None:
+        """Send Telegram to routing connected device."""
+        cemi = CEMIFrame.init_from_telegram(
+            telegram=telegram,
+            code=CEMIMessageCode.L_DATA_IND,
+            src_addr=self.xknx.own_address,
+        )
+        routing_indication = RoutingIndication(cemi=cemi)
+
+        async with self._flow_control.throttle():
+            self._send_knxipframe(KNXIPFrame.init_from_body(routing_indication))
+
+    def _send_knxipframe(self, knxipframe: KNXIPFrame) -> None:
+        """Send KNXIPFrame to connected routing device."""
+        self.udp_transport.send(knxipframe)
+
+    ##################
+    #
+    # INCOMING FRAMES
+    #
+    ##################
+
+    def _handle_frame(
+        self, knxipframe: KNXIPFrame, source: HPAI, _: KNXIPTransport
+    ) -> None:
+        """Handle incoming KNXIPFrames. Callback from internal udp_transport."""
+        if isinstance(knxipframe.body, RoutingIndication):
+            self._handle_routing_indication(knxipframe.body)
+        elif isinstance(knxipframe.body, RoutingBusy):
+            self._flow_control.handle_routing_busy(knxipframe.body)
+        elif isinstance(knxipframe.body, RoutingLostMessage):
+            logger.warning(
+                "RoutingLostMessage received from %s - %s lost messages.",
+                source.ip_addr,
+                knxipframe.body.lost_messages,
+            )
+        else:
+            logger.warning("Service not implemented: %s", knxipframe)
+
+    def _handle_routing_indication(self, routing_indication: RoutingIndication) -> None:
+        """Handle incoming RoutingIndication."""
+        if routing_indication.cemi is None:
+            # Don't handle invalid cemi frames (None)
+            return
+        if routing_indication.cemi.src_addr == self.xknx.own_address:
+            logger.debug("Ignoring own packet")
+            return
+
+        # TODO: is cemi message code L_DATA.req or .con valid for routing? if not maybe warn and ignore
+        asyncio.create_task(self.handle_cemi_frame(routing_indication.cemi))
+
+    async def handle_cemi_frame(self, cemi: CEMIFrame) -> None:
+        """Handle incoming telegram and send responses if applicable (device management)."""
+        telegram = cemi.telegram
+        telegram.direction = TelegramDirection.INCOMING
+
+        if response_tgs := await self.telegram_received_callback(telegram):
+            for response in response_tgs:
+                await self.send_telegram(response)