+++ xknx/knxip/__init__.py
@@ -33,7 +33,9 @@ from .knxip_enum import (
     KNXMedium,
     SearchRequestParameterType,
 )
+from .routing_busy import RoutingBusy
 from .routing_indication import RoutingIndication
+from .routing_lost_message import RoutingLostMessage
 from .search_request import SearchRequest
 from .search_request_extended import SearchRequestExtended
 from .search_response import SearchResponse
@@ -77,7 +79,9 @@ __all__ = [
     "HostProtocol",
     "KNXIPServiceType",
     "KNXMedium",
+    "RoutingBusy",
     "RoutingIndication",
+    "RoutingLostMessage",
     "SearchRequest",
     "SearchRequestExtended",
     "SearchRequestParameterType",